Split log.rs into record and manager files
