+++ src/log/mod.rs
@@ -1,1 +1,2 @@
-pub mod log;
+pub mod manager;
+pub mod record;
