+++ src/tx/transaction.rs
@@ -4,7 +4,8 @@ use std::sync::{Arc, Mutex};
 
 use crate::buffer::BufferManager;
 use crate::file::BlockId;
-use crate::log::log::{LogManager, LogRecord};
+use crate::log::manager::LogManager;
+use crate::log::record::LogRecord;
 
 use super::concurrency::{ConcurrencyManager, LockTable};
